Task #1236: Verklig tid & automatiskt avslut
- Schema: workOrders.actualTimeGroupKey (text), actualDurationManual
  (boolean, default false), index idx_work_orders_actual_time_group;
  new task_events types: auto_completed, actual_time_distributed,
  actual_time_adjusted.
- server/services/active-task-guard.ts: closeOtherActiveWork() enforces
  "one active task at a time" per resource — auto-completes other open
  work_orders (onSiteAt set, completedAt null) and closes other open
  work_entries (endTime null) when a new one starts. Best-effort,
  never throws. Wired into mobile/orders.ts (status -> paborjad) and
  mobile/workSessions.ts (new open time entry).
- server/services/actual-time-distribution.ts: distributeActualTime()
  does a largest-remainder proportional split of a registered total
  across an explicit list of work order ids (klump = client-supplied
  workOrderIds[], no persisted klump table exists yet), weighted by
  estimatedDuration, respecting actualDurationManual-locked rows.
  adjustActualTimeDistribution() is the manual override path, sets
  actualDurationManual=true.
- New routes: POST /api/work-orders/actual-time/distribute and /adjust
  (requirePlanner-gated), POST /api/mobile/orders/actual-time/distribute
  (verifies resource/tenant ownership of all ids).
- Minimal frontend surface: WorkOrderDetailPage shows a "justera" inline
  editor (admin-only) next to "Faktisk tid", posting to the adjust
  endpoint, plus a "(klumpfördelad)" hint when actualTimeGroupKey is set.
- Verified functionally: distribution largest-remainder split sums
  correctly and respects locks; active-task guard auto-closes a prior
  open work order with correct actualDuration/status; schema-drift and
  full tsc typecheck clean (only pre-existing baseline errors in
  untouched files).
- Payroll/invoicing/statistics consumers (kpiRoutes, weekly-report,
  roiRoutes, sla-risk-engine, ai-planner, disruption-service) already
  read actualDuration broadly, so no additional wiring was needed there.
- Out of scope (per task): time-type-registry rule engine (#1237),
  168-hour view, and building a persisted klump/grouping table
  (#1239, future Klumpmotor).

+++ server/routes/mobile/orders.ts
@@ -631,7 +631,7 @@ app.post("/api/mobile/orders/actual-time/distribute", isMobileAuthenticated, asy
   const resourceId = req.mobileResourceId;
   const schema = z.object({
     workOrderIds: z.array(z.string()).min(1),
-    totalActualMinutes: z.number().min(0),
+    totalActualMinutes: z.number().int().min(0),
   });
   const parsed = schema.safeParse(req.body);
   if (!parsed.success) throw new ValidationError(formatZodError(parsed.error).error);
